requirements.txt
Django==4.2.7
djangorestframework==3.14.0
requests==2.31.0
pandas==2.1.1
scikit-learn==1.3.2
numpy==1.26.1
python-dotenv==1.0.0
pyjwt==2.8.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.0
pytest==7.4.3
pytest-django==4.7.0
joblib==1.3.2
matplotlib==3.8.0
seaborn==0.13.0
coverage==7.3.2
black==23.9.1
isort==5.12.0
flake8==6.1.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7